pyproject.toml
[tool.poetry]
name = "goit-python-web-hw-09"
version = "0.1.0"
description = ""
authors = ["lexxai <lexxai@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
pymongo = "^4.5.0"
mongoengine = "^0.27.0"
faker = "^19.10.0"
tdqm = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
